Refactor Options.cmdline() to better handle CLI args and config files when both are present.
Known args, such as --config-file and --no, that are unset are now removed from the arg dict before it is processed so they are not written to a config file later.
Configuration file level options take precedence over command line args.
config_file is no longer written to the configuration file itself. This served no purpose and it was not clear to the user how config_file in both the file and the command line arguments interacted.
It is now possible to specify the subcommand from the config file without specifying it via the command line.