Update modules dealing with kernel parameters to not assume the parameter exists.
Kernel parameters can and do change over time. As such, we cannot assume any given parameter already exists. To address this, these modules have been standardized on iterating over all the parameters.
The particular, unhandled issue that necessitated this change was the removal of tcp_tw_recycle in the 4.12 kernel.